Support packs with multiple files and multiple chart folders
Rather than just copying a Dockerfile and folder called charts it would be handy if we can add extra files and folders to our draft packs.

We've found that we add extra files such as a Jenkinsfile (for CI/CD pipelines) and extra chart folders used for deploying into different environments.